Added validation code and changed estimation code to work with non normalize values

+++ src/optimization/validation/read_data/Read_Coef.cpp
@@ -1,5 +1,5 @@
 //
-// Created by danctorres on 3/7/2023.
+// Created by danctorres on 05/08/2023.
 //
 
 #include "Read_Coef.h"
@@ -8,9 +8,9 @@
 Read_Coef::Read_Coef(const std::string &name_file) : Read(name_file) {
     const std::vector<std::vector<std::string>> data = this->get_data();
     for (int i = 1; i < data.size() - 1; i++){
-        coef_0.push_back(std::stoi(data[i][0]));
-        coef_1.push_back(std::stoi(data[i][1]));
-        coef_2.push_back(std::stoi(data[i][2]));
+        coef_0.push_back(std::stof(data[i][0]));
+        coef_1.push_back(std::stof(data[i][1]));
+        coef_2.push_back(std::stof(data[i][2]));
     }
 }
 
@@ -18,4 +18,3 @@ std::vector<float> Read_Coef::get_coef_eq(const int eq_idx) const {
     std::vector<float> eq_coef = {coef_0[eq_idx], coef_1[eq_idx], coef_2[eq_idx]};
     return eq_coef;
 }
-
